requirements.txt
stable-baselines3==2.0.0
sb3-contrib==2.0.0
gymnasium==0.29.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
torch==2.0.1
pyyaml==6.0
tensorboard==2.13.0
numba==0.57.1
TA-Lib==0.4.26
